Phase 1: codepage tables, the rich-clipboard facade, and a corpus gate
Three agents' work plus fixes for what it turned up.

rclip-codepage — twelve single-byte encodings, written rather than
reused. Six candidate crates were built and run byte-by-byte, and every
one of them silently fills the C1-range holes: encoding_rs and yore
produce byte-identical undefined sets omitting 0x81/0x8D/0x8F/0x90/0x9D
in CP1252, which decode to U+0081 with had_errors == false. For a
clipboard codec, being able to say "this byte is unassigned" is the one
property that could not be given up. encoding_rs also has no CP437 or
CP850 and never will (WHATWG scope), and its for_label("iso-8859-1")
returns windows-1252.

Tables are generated from the Unicode Consortium vendor files, each
pinned by SHA-256 with a mismatch as a hard error, and cross-checked
against CPython's independent codec tables: zero disagreements across
all twelve times 128 entries. decode_byte follows Unicode and reports
undefined; decode_byte_lenient follows Windows. Neither is silent.

Wired behind a default-off codepage feature into the five crates that
had deferred ANSI handling in Phase 0.

rich-clipboard — the facade. Feature-gated per format, with the
write-side fan-out as its own table rather than reusing read_rank: the
two genuinely disagree, and CF_DIBV5 before PNG on Windows is the
concrete case. Reading wants the unambiguous alpha convention, writing
wants what Paint will actually accept. Fidelity is in the type system,
not a comment.

rclip-corpus-tests — one gate over all 176 fixtures plus a leak
scanner. It already earned itself: it caught a live WebKit origin UUID
in a Safari capture on its first run, before anyone read the report.

Fixes for defects the agents found in each other's work:

  - rclip-cf-html did not compile for any 32-bit target. FIELD_MAX was
    a ten-digit usize literal, which is a hard error where usize is 32
    bits. Now u64 with widening comparisons. CI missed it because the
    bare-metal job only built --no-default-features, so nothing behind
    the alloc feature was ever built for a 32-bit target; that job now
    builds with alloc too.
  - cida-child-offset-past-end.bin had aoffset[0] = 16, copied from the
    two-child fixture where the header really is 16 bytes. With cidl=1
    the header ends at 12, so the parent offset pointed four bytes into
    the first item. Its test only called parent().is_ok(), which hands
    back an iterator and succeeds without reading a byte, so the test
    passed while asserting nothing. Fixture corrected and the test now
    walks the parent to completion.
  - two-paths-wide.bin spelled a real first name into a Windows home
    path in a public repo. Replaced with a placeholder at identical
    byte length, and the leak scanner's allowlist entry for it deleted
    rather than kept — the scanner passes on the merits now.

+++ crates/rclip-cf-html/src/serialize.rs
@@ -39,7 +39,12 @@ use crate::{Version, END_FRAGMENT_COMMENT, START_FRAGMENT_COMMENT};
 const FIELD_WIDTH: usize = 10;
 
 /// Largest value a [`FIELD_WIDTH`]-digit field can hold.
-const FIELD_MAX: usize = 9_999_999_999;
+///
+/// `u64` rather than `usize` because ten digits overflow a 32-bit `usize`, and
+/// a literal that does not fit is a hard compile error rather than a warning —
+/// so typing this as `usize` makes the crate simply not build for a 32-bit
+/// target. Comparisons widen the offset instead.
+const FIELD_MAX: u64 = 9_999_999_999;
 
 /// The default document wrapped around a fragment when no context is given.
 const DEFAULT_CONTEXT: (&str, &str) = ("<html><body>", "</body></html>");
@@ -192,7 +197,7 @@ impl<'a> CfHtmlBuilder<'a> {
         }
         let end_html = buf.len();
 
-        if end_html > FIELD_MAX {
+        if end_html as u64 > FIELD_MAX {
             return Err(Error::new(ErrorKind::TooLarge, end_html));
         }
 
@@ -283,7 +288,7 @@ fn reserve(buf: &mut Vec<u8>, key: &str, placeholder: bool) -> Option<usize> {
 /// `field` is `None` for a `-1` line, which has nothing to patch.
 fn patch(buf: &mut [u8], field: Option<usize>, value: usize) -> Result<()> {
     let Some(at) = field else { return Ok(()) };
-    if value > FIELD_MAX {
+    if value as u64 > FIELD_MAX {
         return Err(Error::new(ErrorKind::TooLarge, value));
     }
     let slot = buf
